test(assistant-engine): isolate controller coverage
